+++ client/features/playground/utils/rectHandler.ts
@@ -1,0 +1,4 @@
+export const rectHandler = (current: HTMLDivElement | null) => {
+  const rect = current?.getBoundingClientRect();
+  return rect ? { x: rect.x, y: rect.y, h: rect.height } : { x: 0, y: 0, h: 0 };
+};
